requirements.txt
python-dotenv~=0.15.0
sshpubkeys~=3.3.0
discord.py~=1.6.0